+++ NEWS.md
@@ -29,6 +29,11 @@
 * mkvmerge: Matroska reader: when reading WebM files that use block additions
   with ID 4 (HDR for VP9) an appropriate block addition mapping will be
   created automatically. Part of the implementation of #3511.
+* MKVToolNix GUI: multiplexer: double-clicking on certain columns in the
+  "tracks" view will now toggle them directly (columns "Default track" &
+  "Forced track") or activate the corresponding control (columns "Language",
+  "Name" & "Delay"). Double-clicks on other columns will do what they've done
+  before, too: toggling the "Copy item" state. Implements #3513 & others.
 
 ## Bug fixes